test: add a test to make sure WriteableBitmaps set as Backgrounds invalidate correctly
